+++ Parser/Model/TeamInRank.cs
@@ -7,7 +7,7 @@ namespace Parser.Model
         public int Rank { get; set; }
         public string Name { get; set; }
         public double TotalPoints { get; set; }
-        public int PreviousPoints { get; set; }
+        public double PreviousPoints { get; set; }
         public int MovePosition { get; set; }
         public string FlagUrl { get; set; }
     }
